requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
bidict==0.22.1
frozenlist==1.4.1
h11==0.14.0
idna==3.6
multidict==6.0.4
netifaces==0.10.6
python-engineio==4.8.1
python-socketio==5.10.0
simple-websocket==1.0.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
